# Sieve::factor: overflow fix and early-break optimization