This PR adds a fix for multiplication overflow, with a testcase that triggered it.
It also adds an early break when primes are too big to be a factor, with benchmarks.