@cuviper 2015-09-21
Benchmarks before:

```
test sieve::benches::factor_composite              ... bench:         251 ns/iter (+/- 20)
test sieve::benches::factor_large_prime            ... bench:      75,523 ns/iter (+/- 5,302)
test sieve::benches::factor_medium_prime           ... bench:      11,398 ns/iter (+/- 1,085)
test sieve::benches::factor_over_prime             ... bench:      73,823 ns/iter (+/- 1,561)
test sieve::benches::factor_small_prime            ... bench:         415 ns/iter (+/- 6)
```

After:

```
test sieve::benches::factor_composite              ... bench:         246 ns/iter (+/- 99)
test sieve::benches::factor_large_prime            ... bench:         649 ns/iter (+/- 30)
test sieve::benches::factor_medium_prime           ... bench:         304 ns/iter (+/- 16)
test sieve::benches::factor_over_prime             ... bench:       1,920 ns/iter (+/- 52)
test sieve::benches::factor_small_prime            ... bench:         101 ns/iter (+/- 2)
```