@cuviper 2016-01-23
Any chance of a new release on crates.io?
I've been pointing my code to git just for this speedup... :)